// БР2.2/Зеленков Евгений/labs/lab2/app/src/shared/enums.ts
export enum Role {
  Admin = 'Admin',
  User = 'User'
}

export enum ReservationStatus {
  Pending = 'Pending',
  Confirmed = 'Confirmed',
  Cancelled = 'Cancelled',
  Completed = 'Completed'
}

export enum Cuisine {
  European = 'European',
  Asian = 'Asian',
  Eastern = 'Eastern',
  Caucasian = 'Caucasian',
  American = 'American',
  Vegetarian = 'Vegetarian'
}